regex_TranAnhQuan

+++ Duan1_Nhom3_IT16301/src/View/QuanLiNhanVien.java
@@ -16,6 +16,7 @@ import java.io.File;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.List;
+import java.util.regex.Pattern;
 import javax.swing.ImageIcon;
 import javax.swing.table.DefaultTableModel;
 
@@ -25,6 +26,8 @@ import javax.swing.table.DefaultTableModel;
  */
 public class QuanLiNhanVien extends javax.swing.JFrame {
 
+    private List<NhanVien> list;
+    NhanVienDAO NVDao = new NhanVienDAO();
     /**
      * Creates new form QuanLiNhanVien
      */
@@ -35,6 +38,7 @@ public class QuanLiNhanVien extends javax.swing.JFrame {
         Load();
         setLocationRelativeTo(null);
         rd_Nam.setSelected(true);
+        list = NVDao.select();
 
     }
     NhanVienDAO dao = new NhanVienDAO();
@@ -106,7 +110,12 @@ public class QuanLiNhanVien extends javax.swing.JFrame {
 
     void Delete() {
         try {
+
             String MaNV = txt_MaNV.getText();
+            if (MaNV.equals("")) {
+                DialogHelper.alert(this, "Mời bạn chọn nhân viên");
+                return;
+            }
             dao.delete(MaNV);
             DialogHelper.alert(this, "Xóa Nhân viên Thành Công");
             clear();
@@ -159,7 +168,6 @@ public class QuanLiNhanVien extends javax.swing.JFrame {
 //        Image im = icon1.getImage();
 //        ImageIcon icon = new ImageIcon(im.getScaledInstance(lb_Avatar.getWidth(), lb_Avatar.getHeight(), im.SCALE_SMOOTH));
 //        lb_Avatar.setIcon(icon);
-
         lb_Avatar.setToolTipText(model.getAvatar());
         if (model.getAvatar() != null) {
             ImageIcon icon = ShareHelper.readLogo(model.getAvatar());
@@ -396,11 +404,11 @@ public class QuanLiNhanVien extends javax.swing.JFrame {
             jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel2Layout.createSequentialGroup()
                 .addGap(0, 0, Short.MAX_VALUE)
-                .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(btnFisrt)
-                    .addComponent(btnPrev)
-                    .addComponent(btnNext)
-                    .addComponent(btnLast)))
+                .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(btnFisrt, javax.swing.GroupLayout.Alignment.TRAILING)
+                    .addComponent(btnPrev, javax.swing.GroupLayout.Alignment.TRAILING)
+                    .addComponent(btnNext, javax.swing.GroupLayout.Alignment.TRAILING)
+                    .addComponent(btnLast, javax.swing.GroupLayout.Alignment.TRAILING)))
         );
 
         tbl_NhanVien.setFont(new java.awt.Font("Monospaced", 0, 18)); // NOI18N
@@ -453,23 +461,20 @@ public class QuanLiNhanVien extends javax.swing.JFrame {
                             .addComponent(jLabel1, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
                             .addComponent(jLabel2, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
                             .addComponent(jLabel7, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                        .addGap(24, 24, 24)
                         .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                            .addGroup(jPanel1Layout.createSequentialGroup()
-                                .addGap(24, 24, 24)
-                                .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
-                                    .addComponent(txt_MaNV)
-                                    .addComponent(txt_HoTen)
-                                    .addComponent(txt_user)
-                                    .addComponent(date_chooser, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                                    .addComponent(txt_Phone, javax.swing.GroupLayout.PREFERRED_SIZE, 271, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel1Layout.createSequentialGroup()
-                                .addGap(24, 24, 24)
-                                .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
-                                    .addComponent(txt_DiaChi, javax.swing.GroupLayout.PREFERRED_SIZE, 271, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                    .addGroup(jPanel1Layout.createSequentialGroup()
-                                        .addComponent(rd_Nam)
-                                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                                        .addComponent(rd_Nu, javax.swing.GroupLayout.PREFERRED_SIZE, 199, javax.swing.GroupLayout.PREFERRED_SIZE)))))
+                            .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                                .addComponent(txt_MaNV)
+                                .addComponent(txt_HoTen)
+                                .addComponent(txt_user)
+                                .addComponent(date_chooser, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                .addComponent(txt_Phone, javax.swing.GroupLayout.PREFERRED_SIZE, 271, javax.swing.GroupLayout.PREFERRED_SIZE))
+                            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                                .addComponent(txt_DiaChi, javax.swing.GroupLayout.PREFERRED_SIZE, 271, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addGroup(jPanel1Layout.createSequentialGroup()
+                                    .addComponent(rd_Nam)
+                                    .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                    .addComponent(rd_Nu, javax.swing.GroupLayout.PREFERRED_SIZE, 199, javax.swing.GroupLayout.PREFERRED_SIZE))))
                         .addGap(68, 68, 68)
                         .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
                             .addGroup(jPanel1Layout.createSequentialGroup()
@@ -492,7 +497,7 @@ public class QuanLiNhanVien extends javax.swing.JFrame {
         jPanel1Layout.setVerticalGroup(
             jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(jPanel1Layout.createSequentialGroup()
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                .addContainerGap(13, Short.MAX_VALUE)
                 .addComponent(jLabel9)
                 .addGap(1, 1, 1)
                 .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
@@ -559,7 +564,7 @@ public class QuanLiNhanVien extends javax.swing.JFrame {
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(layout.createSequentialGroup()
                 .addContainerGap()
-                .addComponent(jPanel1, javax.swing.GroupLayout.PREFERRED_SIZE, 679, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addComponent(jPanel1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
                 .addComponent(jButton5))
         );
@@ -572,11 +577,16 @@ public class QuanLiNhanVien extends javax.swing.JFrame {
     }//GEN-LAST:event_lb_AvatarMouseClicked
 
     private void btn_ThemActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btn_ThemActionPerformed
-        Insert();
+         if (Check()==false) {
+             Insert();
+        }
+        
     }//GEN-LAST:event_btn_ThemActionPerformed
 
     private void btn_SuaActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btn_SuaActionPerformed
-        Update();
+       if (Check()==false) {
+               Update();
+        }
     }//GEN-LAST:event_btn_SuaActionPerformed
 
     private void btn_XoaActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btn_XoaActionPerformed
@@ -657,7 +667,53 @@ public class QuanLiNhanVien extends javax.swing.JFrame {
             }
         });
     }
+ public boolean Check() {
+        for (int i = 0; i < list.size(); i++) {
+            Pattern phone = Pattern.compile("^[0-9]{10}$");
+            NhanVien s = list.get(i);
+            String MaNV = s.getMaNV();
+            String user = s.getUsername();
+            try {
+                if (txt_MaNV.getText().equals("") || txt_MaNV.getText().length() > 15) {
+                    DialogHelper.alert(this, "Mã Nhân viên không hợp lệ");
+                    txt_MaNV.requestFocus();
+                    return true;
+                } else if (txt_MaNV.getText().equals(MaNV)) {
+                    DialogHelper.alert(this, "Mã Nhân viên đã tồn tại");
+                    txt_MaNV.requestFocus();
+                    return true;
+                } else if (txt_HoTen.getText().equals("")) {
+                    DialogHelper.alert(this, "Tên Nhân viên không được trống");
+                    txt_HoTen.requestFocus();
+                    return true;
 
+                } else if (!txt_user.getText().equals(user)) {
+                    DialogHelper.alert(this, "Tên Username  không tồn tại");
+                    txt_user.requestFocus();
+                    return true;
+                } else if (txt_Phone.getText().equals("")) {
+                    DialogHelper.alert(this, "Số Điện Thoại không được trống");
+                    txt_Phone.requestFocus();
+                    return true;
+                } else if (!phone.matcher(txt_Phone.getText()).find()) {
+                    DialogHelper.alert(this, "Số Điện Thoại không hợp lệ");
+                    txt_Phone.requestFocus();
+                    return true;
+                } else if (txt_DiaChi.getText().equals("")) {
+                    DialogHelper.alert(this, "ĐịaChỉ không được trống");
+                    txt_DiaChi.requestFocus();
+                    return true;
+                } else if (lb_Avatar.getIcon()==null) {
+                    DialogHelper.alert(this, " Ảnh không được trống");
+                    return true;
+                }
+
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
+        }
+        return false;
+    }
     // Variables declaration - do not modify//GEN-BEGIN:variables
     private javax.swing.JButton btnFisrt;
     private javax.swing.JButton btnLast;
